Un-break doctests in counter.py.
These test broke as we're now reporting back if packets are making it to
the destination.  These tests are a bit suspicious to begin with, as
they're actually sending data out on the wire every time you run them,
which isn't quite what unittests should be doing.

The fix implemented here uses nose's ability to run module-level
setup/teardown functions to be able to mock out `Connection.send()`.