Mock `time.time()` calls in Timer tests.
The timer tests required that running an empty function would take less
than a millisecond.  If these tests are run on a heavily loaded system,
this might not always be the case, so the test can fail for reasons not
relating to code correctness.  Thus, we mock out `time.time()` to always
return the same value, to remove the real runtime dependance from the
code.